ensure to wait for app.js loaded before "eval" sfz loader function().
Before this change, there was an error:

> undefined:1:24: CONSOLE ERROR ReferenceError: Can't find variable: selectInstrumentFromSfz

It was because SFZ filename was passed back to Web UI before it completes
loading app.js. Now it awaits until it's loaded, using `await sleep()` trick.

The sleeper function is defined in place here because the exact problem
here is that JS on the web side is NOT loaded yet (i.e. it must be
defined at cpp side).