Now LV2UI can restore GUI at load time, with solid initialized callback.
There is a new callback from Web UI to lv2ui 'I' message to indicate that
it finished loading the JS UI code.
It is then used to call patch:get() to sfizz audio engine to notify sfzfile,
which in turn call SFZ UI loader on the Web UI.

Therefore, there is no need to wait for loading app.js (and namely
`loadInstrumentFromSfz()` function) at C code anymore. So it is simplified.

This change also renamed a bunch of callbacks (prefixing "Aria2Web") to
avoid possible conflict with other JS code.